fix: model page switch marketplace (#17147)
